Refactoring ROCA/VCA classes + optimized ROCA
Moved implementation classes to "algorithms" package, instead of
"datastructures".

Observation table for ROCA now uses a tree to internally store the
observations. This reduces the number of asked membership and counter
value queries. It also improves the time needed to determine whether a
word is in the known prefix of the target language.