New learning algorithm for ROCAs
Approx sets are now define over the whole table (so, short and long
prefixes are handled the same way).
The table must be closed, Sigma consistent, and bottom consistent. In
that case, the Approx sets define an equivalence relation.

Also rewrote some parts to have cleaner and more readable code.